+++ gd-gltg-bmp/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GLGVO-724](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-724)
 - Add top ten practices by area treated to BMP results
   [GEOD-1364](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1364)
+- Export results as PDF
+  [GEOD-1366](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1366)
+  [GH-42](https://github.com/geostreams/geodashboard/issues/42)
 
 ## [3.6.0] - 2020-12-14